+++ changelog.md
@@ -16,6 +16,7 @@ Codion Change Log
 - EntityEditModel, add/removeListener methods replaced with methods that return the underlying event observer.
 ### is.codion.swing.common.model
 - FilteredTableSortModel, add/removeListener methods replaced with methods that return the underlying event observer.
+- FilteredTableSearchModel, add/removeListener methods replaced with methods that return the underlying event observer.
 
 ## 0.17.34
 ### is.codion.common.core